backend/requirements.txt
# Core
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic>=2.10
pydantic-settings>=2.7

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14

# Auth
PyJWT[cryptography]>=2.10
pwdlib[argon2]>=0.2

# Azure
azure-ai-documentintelligence>=1.0.2
openai>=1.60
azure-identity>=1.19
azure-storage-blob>=12.24

# Document conversion
openpyxl>=3.1
python-docx>=1.1
Pillow>=11.0

# Utilities
httpx>=0.28
python-multipart>=0.0.18
loguru>=0.7
orjson>=3.10
sqlparse>=0.5

# Testing
pytest>=8.3
pytest-asyncio>=0.24
